Delete utils + improve install
