+++ data/wp/wp-content/mu-plugins/EPFL-news.php
@@ -8,12 +8,60 @@
 
 declare(strict_types=1);
 
-require_once('utils.php');
-
-use utils\Utils as Utils;
-
 define("NEWS_API_URL", "https://actu.epfl.ch/api/v1/channels/");
 define("NEWS_API_URL_IFRAME", "https://actu.epfl.ch/webservice_iframe/");
+
+Class NewsUtils
+{
+    public static function debug($var) {
+        print "<pre>";
+        var_dump($var);
+        print "</pre>";
+    }
+
+    /**
+     * This allow to insert anchor before the element
+     *   i.e. '<a name="' . $ws->get_anchor($item->title) . '"></a>';
+     * and also to get the item link in case it's not provided by the API.
+     * e.g. https://actu.epfl.ch/news/a-12-million-franc-donation-to-create-a-center-for/
+     */
+    public static function get_anchor(string $title): string {
+
+        $unwanted_array = array(    'Š'=>'S', 'š'=>'s', 'Ž'=>'Z', 'ž'=>'z', 'À'=>'A', 'Á'=>'A', 'Â'=>'A', 'Ã'=>'A', 'Ä'=>'A', 'Å'=>'A', 'Æ'=>'A', 'Ç'=>'C', 'È'=>'E', 'É'=>'E',
+                                    'Ê'=>'E', 'Ë'=>'E', 'Ì'=>'I', 'Í'=>'I', 'Î'=>'I', 'Ï'=>'I', 'Ñ'=>'N', 'Ò'=>'O', 'Ó'=>'O', 'Ô'=>'O', 'Õ'=>'O', 'Ö'=>'O', 'Ø'=>'O', 'Ù'=>'U',
+                                    'Ú'=>'U', 'Û'=>'U', 'Ü'=>'U', 'Ý'=>'Y', 'Þ'=>'B', 'ß'=>'Ss', 'à'=>'a', 'á'=>'a', 'â'=>'a', 'ã'=>'a', 'ä'=>'a', 'å'=>'a', 'æ'=>'a', 'ç'=>'c',
+                                    'è'=>'e', 'é'=>'e', 'ê'=>'e', 'ë'=>'e', 'ì'=>'i', 'í'=>'i', 'î'=>'i', 'ï'=>'i', 'ð'=>'o', 'ñ'=>'n', 'ò'=>'o', 'ó'=>'o', 'ô'=>'o', 'õ'=>'o',
+                                    'ö'=>'o', 'ø'=>'o', 'ù'=>'u', 'ú'=>'u', 'û'=>'u', 'ý'=>'y', 'þ'=>'b', 'ÿ'=>'y' );
+
+        $title = strtr( $title, $unwanted_array );
+        $title = str_replace(" ", "-", $title);
+        $title = str_replace("'", "-", $title);
+        $title = strtolower($title);
+        $title = substr($title, 0, 50);
+
+        return $title;
+    }
+
+    /**
+     * Call API
+     * @param url  : the fetchable url
+     * @param args : array('timeout' => 10), see https://codex.wordpress.org/Function_Reference/wp_remote_get
+     * @return decoded JSON data
+     */
+    public static function get_items(string $url) {
+
+        $response = wp_remote_get($url);
+
+        if (is_array($response)) {
+                $header = $response['headers']; // array of http header lines
+                $data = $response['body']; // use the content
+                if ( $header["content-type"] === "application/json" ) {
+                        $items = json_decode($data);
+                        return $items;
+                }
+        }
+    }
+}
 
 /**
  * Template text only
@@ -58,13 +106,13 @@ function epfl_news_template_fac_with_4_news($news, bool $stickers): string
 	foreach ($news->results as $item) {
 		$html .= '<article class="post">';
 		$html .= '  <figure class="post-thumbnail">';
-		$html .= '    <a href="https://actu.epfl.ch/news/' . Utils::get_anchor($item->title) . '">';
+		$html .= '    <a href="https://actu.epfl.ch/news/' . NewsUtils::get_anchor($item->title) . '">';
 		$html .= '      <img src="' . $item->visual_url . '" title="'.$item->title.'">';
 		$html .= '    </a>';
 		$html .= '  </figure>';
 		$html .= '  <header class="entry-header">';
 		$html .= '    <h2 class="entry-title">';
-		$html .= '      <a href="https://actu.epfl.ch/news/' . Utils::get_anchor($item->title) . '">';
+		$html .= '      <a href="https://actu.epfl.ch/news/' . NewsUtils::get_anchor($item->title) . '">';
 		$html .= $item->title;
 		$html .= '      </a>';
 		$html .= '    </h2>';
@@ -96,13 +144,13 @@ function epfl_news_template_labo_with_5_news($news, bool $stickers): string
 		$html .= '<article class="post">';
 		$html .= '  <header class="entry-header">';
 		$html .= '    <h2 class="entry-title">';
-		$html .= '      <a href="https://actu.epfl.ch/news/' . Utils::get_anchor($item->title) . '">';
+		$html .= '      <a href="https://actu.epfl.ch/news/' . NewsUtils::get_anchor($item->title) . '">';
 		$html .= $item->title;
 		$html .= '      </a>';
 		$html .= '    </h2>';
 		$html .= '  </header>';
 		$html .= '  <figure class="post-thumbnail">';
-		$html .= '    <a href="https://actu.epfl.ch/news/' . Utils::get_anchor($item->title) . '">';
+		$html .= '    <a href="https://actu.epfl.ch/news/' . NewsUtils::get_anchor($item->title) . '">';
 		$html .= '      <img src="' . $item->visual_url . '" title="'.$item->title.'">';
 		$html .= '    </a>';
 		$html .= '  </figure>';
@@ -134,9 +182,9 @@ function epfl_news_template_labo_with_3_news($news, bool $stickers): string
 	    // print fr and en category
 	    if ($stickers == TRUE) {
             if ($item->lang === "fr") {
-                Utils::debug($item->category->fr_label);
+                NewsUtils::debug($item->category->fr_label);
             } elseif ($item->lang === "en") {
-                Utils::debug($item->category->en_label);
+                NewsUtils::debug($item->category->en_label);
             }
         }
         $publish_date = new DateTime($item->publish_date);
@@ -145,13 +193,13 @@ function epfl_news_template_labo_with_3_news($news, bool $stickers): string
 		$html .= '<article class="post">';
 		$html .= '  <header class="entry-header">';
 		$html .= '    <h2 class="entry-title">';
-		$html .= '      <a href="https://actu.epfl.ch/news/' . Utils::get_anchor($item->title) . '">';
+		$html .= '      <a href="https://actu.epfl.ch/news/' . NewsUtils::get_anchor($item->title) . '">';
 		$html .= $item->title;
 		$html .= '      </a>';
 		$html .= '    </h2>';
 		$html .= '  </header>';
 		$html .= '  <figure class="post-thumbnail">';
-		$html .= '    <a href="https://actu.epfl.ch/news/' . Utils::get_anchor($item->title) . '">';
+		$html .= '    <a href="https://actu.epfl.ch/news/' . NewsUtils::get_anchor($item->title) . '">';
 		$html .= '      <img src="' . $item->visual_url . '" title="'.$item->title.'">';
 		$html .= '    </a>';
 		$html .= '  </figure>';
@@ -204,7 +252,7 @@ function epfl_news_built_html_pagination_template(string $channel, string $lang)
 
     // call API to get the name of channel
     $url = NEWS_API_URL . $channel;
-    $channel = Utils::get_items($url);
+    $channel = NewsUtils::get_items($url);
 
     $url = NEWS_API_URL_IFRAME . $channel->name . "/" . $lang . "/nosticker";
 
@@ -303,7 +351,7 @@ function epfl_news_check_required_parameters(string $channel, string $lang): boo
 
     // check that the channel exists
     $url = NEWS_API_URL . $channel;
-    $channel_response = Utils::get_items($url);
+    $channel_response = NewsUtils::get_items($url);
     if ($channel_response->detail === "Not found.") {
         return FALSE;
     }
@@ -354,7 +402,7 @@ function epfl_news_process_shortcode(
             $themes
         );
 
-        $actus = Utils::get_items($url);
+        $actus = NewsUtils::get_items($url);
         return epfl_news_build_html($actus, $template, $stickers);
 }
