--- data/wp/wp-content/mu-plugins/utils.php
@@ -1,65 +1,0 @@
-<?php
-
-/**
- * File Name: utils.php
- * Description: it provides methods used by news and memento shortcodes
- * @version: 1.0
- * @copyright: Copyright (c) 2017 Ecole Polytechnique Federale de Lausanne, Switzerland
- */
-
-declare(strict_types=1);
-
-namespace utils;
-
-Class Utils
-{
-    public static function debug($var) {
-        print "<pre>";
-        var_dump($var);
-        print "</pre>";
-    }
-
-    /**
-     * This allow to insert anchor before the element
-     *   i.e. '<a name="' . $ws->get_anchor($item->title) . '"></a>';
-     * and also to get the item link in case it's not provided by the API.
-     * e.g. https://actu.epfl.ch/news/a-12-million-franc-donation-to-create-a-center-for/
-     */
-    public static function get_anchor(string $title): string {
-
-        $unwanted_array = array(    'Š'=>'S', 'š'=>'s', 'Ž'=>'Z', 'ž'=>'z', 'À'=>'A', 'Á'=>'A', 'Â'=>'A', 'Ã'=>'A', 'Ä'=>'A', 'Å'=>'A', 'Æ'=>'A', 'Ç'=>'C', 'È'=>'E', 'É'=>'E',
-                                    'Ê'=>'E', 'Ë'=>'E', 'Ì'=>'I', 'Í'=>'I', 'Î'=>'I', 'Ï'=>'I', 'Ñ'=>'N', 'Ò'=>'O', 'Ó'=>'O', 'Ô'=>'O', 'Õ'=>'O', 'Ö'=>'O', 'Ø'=>'O', 'Ù'=>'U',
-                                    'Ú'=>'U', 'Û'=>'U', 'Ü'=>'U', 'Ý'=>'Y', 'Þ'=>'B', 'ß'=>'Ss', 'à'=>'a', 'á'=>'a', 'â'=>'a', 'ã'=>'a', 'ä'=>'a', 'å'=>'a', 'æ'=>'a', 'ç'=>'c',
-                                    'è'=>'e', 'é'=>'e', 'ê'=>'e', 'ë'=>'e', 'ì'=>'i', 'í'=>'i', 'î'=>'i', 'ï'=>'i', 'ð'=>'o', 'ñ'=>'n', 'ò'=>'o', 'ó'=>'o', 'ô'=>'o', 'õ'=>'o',
-                                    'ö'=>'o', 'ø'=>'o', 'ù'=>'u', 'ú'=>'u', 'û'=>'u', 'ý'=>'y', 'þ'=>'b', 'ÿ'=>'y' );
-
-        $title = strtr( $title, $unwanted_array );
-        $title = str_replace(" ", "-", $title);
-        $title = str_replace("'", "-", $title);
-        $title = strtolower($title);
-        $title = substr($title, 0, 50);
-
-        return $title;
-    }
-
-    /**
-     * Call API
-     * @param url  : the fetchable url
-     * @param args : array('timeout' => 10), see https://codex.wordpress.org/Function_Reference/wp_remote_get
-     * @return decoded JSON data
-     */
-    public static function get_items(string $url) {
-
-        $response = wp_remote_get($url);
-
-        if (is_array($response)) {
-                $header = $response['headers']; // array of http header lines
-                $data = $response['body']; // use the content
-                if ( $header["content-type"] === "application/json" ) {
-                        $items = json_decode($data);
-                        return $items;
-                }
-        }
-    }
-}
-?>
