+++ src/wordpress/generator.py
@@ -371,6 +371,8 @@ class WPGenerator:
         WPMuPluginConfig(self.wp_site, "EPFL-SC-people.php").install()
         WPMuPluginConfig(self.wp_site, "EPFL-map.php").install()
         WPMuPluginConfig(self.wp_site, "EPFL-scheduler.php").install()
+        WPMuPluginConfig(self.wp_site, "EPFL-events.php").install()
+        WPMuPluginConfig(self.wp_site, "EPFL-news.php").install()
 
         if self.wp_config.installs_locked:
             WPMuPluginConfig(self.wp_site, "EPFL_installs_locked.php").install()
